+++ src/js/genres.js
@@ -1,4 +1,4 @@
-genres: [
+export const genres = [
   {
     id: 28,
     name: 'Action',
